Fix 'argument': conversion from 'int' to 'float compiler warning

+++ GaussianFilter.hpp
@@ -81,7 +81,7 @@ private:
         return v;
     }
 
-    float gaussian(float sigma, float value)
+    float gaussian(float sigma, int value)
     {
         return (1.f / std::sqrtf(2.f * M_PI * sigma) * std::expf(-(value * value) / (2.f * sigma * sigma)));
     }
